Make list scrolling and autoload working
Use ItemKeyedDatasource instead of positional data source since
with positional data source we need to know the number of items
in advance, i.e., total number of items must be known in advance
when using positional data source.
With ItemKeyed data source we can just keep on adding unique page
keys and we only stop when the list it empty.
When the end of list is reached it's better to just invalidate the
data source.